Cleanup some of the utils and add comments
